Rename composable to useCourtType.ts
RISDEV-3775
