+++ frontend/src/components/DocumentUnitCategories.vue
@@ -11,7 +11,7 @@ import FlexContainer from "@/components/FlexContainer.vue"
 import FlexItem from "@/components/FlexItem.vue"
 import { ValidationError } from "@/components/input/types"
 import PreviousDecisions from "@/components/PreviousDecisions.vue"
-import { useProvideCourtType } from "@/composables/useProvideCourtType"
+import { useProvideCourtType } from "@/composables/useCourtType"
 import useQuery from "@/composables/useQueryFromRoute"
 import { useScrollToHash } from "@/composables/useScrollToHash"
 import DocumentUnit, { Texts, CoreData } from "@/domain/documentUnit"
